Test-PendingReboot.ps1: Allow for truly local box use (#28)
* Update Test-PendingReboot.ps1

Updated Test-PendingReboot.ps1 to allow for truly local box use. Under the prior method, even on a local box, a remote session was opened and that was incompatible with our environment.

* Update Test-PendingReboot.ps1

Removed some personal debug code from what I wanted in the Pull Request.

* Update Test-PendingReboot.ps1

Updated Test-PendingReboot.ps1 to allow for truly local box use. Under the prior method, even on a local box, a remote session was opened and that was incompatible with our environment.